+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased]
+### Fixed
+- Increased size of most database char columns to default size (255)
+
 ## [0.1.0] – 2024-04-18
 
 ### Added